backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.11
geopandas==1.0.1
shapely==2.0.6
requests==2.32.3
python-dotenv==1.0.1
openai==1.51.0
pandas==2.2.3
numpy==2.1.2
pyproj==3.7.0
gtfs-realtime-bindings==1.0.0
sqlalchemy==2.0.36
GeoAlchemy2==0.15.2
